Test ordering of singular values
Interestingly, the singular values just happen to be
ordered already for the tall and short and matrices,
while they are not ordered correctly
for the square example. Before we implement the
sorting of the matrices, we at least want these tests
to fail, so we must construct new matrix examples
for which the singular values hopefully come out
unsorted.