Add SVD benchmarks
Since the benchmarks test module-internal methods,
the benchmarks can not reside in the benches/ subdirectory.
However, since the benchmarks must be run on nightly, it
involves some complication. Hence, in order to run these
benchmarks, you must add the 'nightly' feature to the compilation,
like so:

`cargo bench --features nightly`