Sort singular values (#46)
* Use SVD definition to verify value/vector triplets

* Clean up SVD tests

* Test ordering of singular values

Interestingly, the singular values just happen to be
ordered already for the tall and short and matrices,
while they are not ordered correctly
for the square example. Before we implement the
sorting of the matrices, we at least want these tests
to fail, so we must construct new matrix examples
for which the singular values hopefully come out
unsorted.

* Add tests for sorted singular values

Before implementing the sort functionality,
it is useful to have some tests that fail because
the singular values are not sorted. However,
it appears that the computation of the SVD
for these examples does not even (currently) finish!

* Fix SVD non-termination issue

Thanks to @AtheMathmo for figuring out the issue.
See https://github.com/AtheMathmo/rulinalg/pull/46 for
a discussion of the problem.

* Correct signs of singular values in SVD

* Rewrite correct_svd_signs to idiomatic style

* Enforce ordering of singular values in SVD

* Clarify SVD documentation

* Implement temporary Epsilon trait

The idea is to remove this once epsilon() lands
in the next version of the num crate.

* Add SVD benchmarks

Since the benchmarks test module-internal methods,
the benchmarks can not reside in the benches/ subdirectory.
However, since the benchmarks must be run on nightly, it
involves some complication. Hence, in order to run these
benchmarks, you must add the 'nightly' feature to the compilation,
like so:

`cargo bench --features nightly`

* Remove 'nightly' benchmarking of internal modules

Instead just bench the public .svd() method along
with the other benchmarks.

* Correct SVD dimensions in documentation