Return 0 from det() instead of LUP error
LUP factorization will return an error whenever one of the diagonals of the U
matrix is == 0. This generally makes sense, because such a matrix is
singular and there will probably be an error down the line anyway. However, the
determinant of this matrix is known, and is 0, because the determinant is
calculated as the product of the diagonal of the triangular matrix.

I believe that it is more general to return the correct result for the
determinant (0) rather than raising an error. Let the end-user decide what to
do about this determinant of 0.