Optimized slice iterators for contiguous data
Added behaviour to the SliceIter and SliceIterMut structs to directly
iterate over the data when it is contiguous in memory. When the data is
not contiguous, the iterators still need to keep track of the row
stride, as well as the position of the iterator in both the column and
the row.