Generalize HouseholderComposition with offset
The composition of Householder matrices for
QR decomposition and Upper Hessenberg decomposition
is very similar. We can exploit this to reuse virtually
all code we have written for HouseholderComposition in the
QR decomposition to implement the Hessenberg decomposition.
We do this by adding a "subdiagonal" parameter to
HouseholderComposition which controls the
subdiagonal to pull Householder reflectors from.